IPAd: add skeleton with HTTPs, PCSC and SGP.32 ASN.1 parser/encoder
In this initial commit we add a skeleton that already contains a libcurl
based HTTPs client and a libpcsclite based smartcard reader support. The
skeleton also contains an ASN.1 parser/encoder. This encoder/decoder
still needs some minor fixups but it is already useable.

in main.c one finds a number of testme_ functions that can be used to
try the already existing features.