euicc: open logical channel, select ISD-R
So far we do not support communication on a logical channel. However, in
practice this would be necessary so that we can communicate to the ISD-R
without disturbing other ongoing eUICC communication. We also do not
select the ISD-R yet. This patch fixes both issues. The user can now
configure a logical channel from 0-3 and the ISD-R is selected on that
channel.