utils: make ipa_asn1c_consume_bytes_cb easier to debug
We should put an assert that gets set of in case the ipa_buf has an
illegal length value (len)

We should also print the data that was emitted by the the ASN.1 encoder.
In case we have problems during encoding we get an estimate how far the
encoder came and possibly locate the field that caused the trouble.