utils: add IPA_ALLOC_ZERO and IPA_ALLOC_N_ZERO
The IPA_ALLOC and IPA_ALLOC_N which we normally use are just using
malloc(). When we allocate some memory using those macros we usually
assert the resulting pointer to make sure the allocation was successful.
We also usually zero out the allocated memory. Let's add a _ZERO variant
for IPA_ALLOC and IPA_ALLOC_N that does this for us. Then we can always
be sure that the allocated memory is present and zeroed out.