cmn_mtl_auth_proc: add new module for the common mutual auth procedure
The Common Mutual Authentication Procedure basically starts a new
session between the eUICC and the RSP server. It is one of the core
procedures.

This implementation lacks the function call to "ES10b.CancelSession" in
case the "ES11.AuthenticateClient" function fails (last step). This
missing part will be added as soon as we have an implementation for the
Common Cancel Session Procedure ready.