esipa_get_eim_pkg: rework implementation
The current implementation of esip_get_eip_pkg still uses the method
where we take all asn.1 structs and copy the contents into an own
struct. We have changed the concept some time ago and decided that it is
easier to use the asn1c generated structs directly, so lets rework this
implementation accordingly. This happens in preparation for implementing
the eUICC package download and execution procedure.