libipa: implement eUICC package download and execute procedure (WIP)
The package download and execute procedure transfers a package containing
ECO and PSMO elements to the eUICC. The eUICC executes those elements and
provides a result and a notification list, which must be fetched from the
eUICC and transfered back to the eIM.

The implementation is nearly complete, but the acknowledgement of the the
notifications is not done correctly yet. The implementation should call
the ES10b functin RemoveNotificationFromList once but there may be multiple
notifications (not just one) that have to be removed.