es10b: rename module es10b.c to es10x.c
The modiule es10b.c should be renamed to es10x.c since it contains the
helper functions, define constants and macros that relate to all ES10
variants, not only to ES10b. This also means that we have to rename the
prefix of the contained functions, macros and define constants.
Unfortunately this also affects all es10x modules we have so far.