utils: add a function to check for tags in a tagList.
Some of the protocol structs (asn.1) use a so called tagList, which is
just an array that contains BER-TLV tags. Such lists are used to list
tags of interest when requesting data for example. To make things
easier, lets add a function that allows us to check for a specific tag
in a tagList. It should be sufficient to be able to process tagLists
with tags of up to 2 byte length.