es10b_retr_notif_from_list: improve API and SGP32 compatibility
The API of the RetrieveNotificationFromList function is a bit clumsy to
use. Lets put the searchCriteria as a field in the request struct and ad
another searchCriteria field where the user can provide the search
criteria in SGP32 format. (the function will convert internally)

As a result we get a notificationList. This notification list also
appears in SGP32 format in some structs where we want to use it. So
let's offer the notificationList in SGP32 format as well.