libipa: use only struct instead of _t types for declrarations
In most places we use struct to declare parameter, members or variables.
But there are also a lot of cases where we just use the _t name of that
struct to declare it without writing struct in front of the name. Both
styles are technically the same, but let's not mix styles.