utils: fix buffer overflow
At the end of a tag list we may run into a buffer overflow since we
access the array before checking the number of remaining bytes. We also
should make sure that we do not interpret incomplete two byte tags as
single byte tags when we hit the end of the array.