ipad: allow the configuration of fallback values for EimConfigurationData
The ES10x function GetEimConfigurationData is not available on consumer
eUICCs. Unfortunately this function must be called on startup to
determine the eimId and the eimFqdn. This means that the IPAd would not
even start when we try to use a consumer eUICC with it. Unfortunately
consumer eUICCs are the only physical eUICCs we currently have available
for testing. So in order to be able to make at least some experiments
with consumer eUICCs, lets ad a (temporary) solution that allows us to
define fallback values for eimId and eimFqdn that we can use to continue
when GetEimConfigurationData fails.

Related: SYS#6742