proc_cmn_mtl_auth: strip unwanted TLV envelopes
The fields serverSignature1 and euiccCiPKIdToBeUsed are wrapped in a TLV
envelope when they are received from the eIM. However, the eUICC expects
the content of the TLV envelope only. This means the envelope must be
stripped before wie put them into the structure that is sent to the
eUICC. In any case we must be careful that we do not destroy field
contents in case the envelope is not present. So lets a function that
parses the tag and the length field properly and then chips it.