utils: add new function to properly parse BTLV headers
Even though the ASN.1 decoder does most of the work there are some
situations where we have to look into a TLV encoded string ourselves.
For this we add a new function ipa_parse_btlv_hdr. This function is also
immediately useful since we can use it to fix ipa_strip_tlv_envelope,
which also does a little bit of TLV parsing, but is not 100% correct.