libipa: first step towards an IoT eUICC emulation
There is already a workaround present that makes it possible to use
consumer eUICCs. However, the solution is not very elegant and since we
have decieded to officially support consumer eUICCs a better solution is
needed. This patch removes the workaround and replaces it with code that
we intend to develop into a complete IoT eUICC emulation.

At the moment this only works for GetEimConfigurationData. The API user
provides an eIM configuration as a BER encoded blob. Whenever.
GetEimConfigurationData is called, the eUICC is not queried. Instead the
BER blob is used and presented to the caller as if it came from the
eUICC.

Related: SYS#6742
Related: SYS#6690