libipa/utils: dynamic buffers for ipa_asn1c_consume_bytes_cb
At the moment we pass a pre-allocated buffer of a fixed size to
ipa_asn1c_consume_bytes_cb. Unfortunately it is not always easy to
estimate in advance how large the encoded data will be we always must
pre-allocate a maximum amount of memory. This ist wasteful since most of
the time we deal with small messages. Let's start with a small buffer
and realloc it when it is exhausted.