http: do not make a new TCP connection for each ESipa request
At the moment each ESipa request towards the eIM means that a new TCP
connection is created. The connection ends when the request is done.
This makes it not only very difficult for the eIM to determine to which
"session" a request belongs. It also puts unnecessary strain on the eIM
server.

SGP.32 also suggests to keep a connection througout multiple requests.
e.g. In Figure 5 we see a single secure connection establishment
followed by multiple ESipa requests. It even should be possible to
perform multiple GetEimPackageRequests in a single connection.