es10b_load_euicc_pkg: begin with adding support for PSMOs on consumer euiccs
PSMOs and ECOs are normally executed directly on the eUICC. However,
in consomuer eUICC this mechanism is not implemented. Here we must look
at each PSMO or ECO and execute an apropriate consumer eUICC function.
In this patch we add support for the "enable" PSMO, which is needed to
enable a profile remotely. Anything else is not implemented yet and will
be added with future patches.