libipa: rework handling of non-volatile states (IoT eUICC emulation)
The eUICC emulation requires the handling of some non volatile state. At
the moment this is in particular the eIM configuration, that is normally
stored directly on the IoT eUICC. However, since consumer eUICCs have no
storage for the eIM configuration, we need to store this data somewhere
else.

The current way to setup the IoT eUICC emulation is not very intuitive.
Let's replace the current setup with some easier to use API where the
API user is not bothered with unnecessary details.