ipa: do euicc package execution in two steps
This is a preparation to make the Generic Package Download and Execution
procedure more rubust. The problem is that when the package is loaded
and the PSMOs are executed, the profile might have changed. This can
cause a short interruption of the IP traffic. When we split the
proecedure in two seteps, we are able to communicate the result on the
re-established IP connection.