main: there is no technical reason for a delay between poll cycles
At the moment the delay is not necessary since we exit when no more eIM
packages are available. This means we will not run in an infinite loop
where we waste CPU cycles with useles polls. However, this may change in
the future and a delay may become necessary again.