proc_direct_prfle_dwnld: fix terminology
The procedure we implement is the "Indirect Profile Download" procedure
and not the "Direct Profile Download" procedure. (Those two procedures
are easy to confuse since they almost look identical. The only fundamental
difference is that a "Direct Profile Download" is negotiated between SMDP+
and IPAd directly, circumventing the eIM)

Related: SYS#6564