es10b_euicc_mem_rst: ensure eIM configuration is deleted on memory reset
To delete the eIM configuration on memory reset, SGP.32 suggests to set
bit 3 of the reset options. Since we delete the eIM configuration in
IoT eUICC emulation mode, we should do the same with real IoT eUICCs.

(Unfortunately the spec is a bit ambigious about the usage of bit 3, see
TODO comment)

Related: SYS#6564