es10b_retr_notif_from_lst: fix compatibility with real IoT eUICCs
the current implementation of RetrieveNotificationsList is incompatible
with real IoT eUICCs and the API is unnecessary complicated as well.
Let's optimize the API so that the API user always gets an
SGP32_RetrieveNotificationsListResponse as result and let's re-arange
the code so that in case a consumer eUICC is used, the
SGP32_RetrieveNotificationsListResponse is derived from the SGP.22
RetrieveNotificationsListResponse.

Related: SYS#6690