proc_eim_pkg_retr: process each eIM package within a new HTTP connection
In its current state, the code does not actively close the connection.
The API user is expected to make a concious decision and then call
ipa_close() to close the eIM connection. This is not necessary, since
the IPAd code can also manage the connection autonomously, so let's
ensure that each eIM package is processed within a fresh connection.

Related: SYS#6564