proc_prfle_inst: fix failure behavior
When the profile download fails, the failure mode may be that the
eUICC aborts the installation and a notification is sent, just as
if the download were successful. In this case we must not execute
the cancelSession procdedure (it would fail since the session
has already ended normally with the notification). We also do not
check the return value of ipa_bpp_segments_encode, which is a
potential null pointer dereference. We also should check if we got
a valid sequence number when removing the notification.

Related: SYS#6564