es10b_euicc_mem_rst: fix implementation
The eUICC memory rest implementation is incomplete and does not work.
This patch completes the implementation and fixes the remaining problems.
Unfortunately the implementation could only be tested with a consumer
eUICC due to the lack of samples (the only IoT eUICC we have has a bug
in the memory rest that was also confirmed by the manufacturer).

Related: SYS#6564